Create README.md
SOLIDITY SUBMISSION
submission for solidity project for metacrafter. This is a contract which adds value to address and detect's it it also stop detecting balance if detecting amount is more than balance.

DESCRIPTION
This program is written in solidity a programming language which is used to write smart contract in Ethereum a leading in blockchain technology, this contract is a simple set of code which is used to add ether to your account and detect it and if the detuction amount is more than avilable it does not detect the amount, other than this it shows the avilable amount in our address.
RUN THE CODE
// SPDX-License-Identifier: MIT
pragma solidity 0.8.18;

contract MyToken {

    // public variables here
    string public tokenName="META";
    string public tokenAbbrv ="MTA";
    uint public totalSupply =0;
    // mapping variable here
    mapping (address => uint) public balances;
    // mint function
    function mint(address _address, uint _value) public {
        totalSupply += _value;
        balances[_address] += _value;
    }
    // burn function
    function burn (address _address,uint _value) public {
        if(balances[_address]>= _value){
        totalSupply -= _value;
        balances[_address] -= _value;
        }
    }
}